requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.2
Django==4.0.3
djangorestframework==3.13.1
paramiko==2.10.3
pycparser==2.21
PyNaCl==1.5.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
